+++ lib/src/functions/select_file.dart
@@ -1,11 +1,14 @@
 import 'package:file_picker/file_picker.dart';
+import 'package:image_picker/image_picker.dart';
+import 'package:path/path.dart' as xpath;
 
 /// select file from  file picker
 /// [maxSize] is in MB
-Future<({String? path, String? name, String? error})> selectFileAsync(
-    {int? maxSize,
-    List<String>? allowedExtensions,
-    String? changeFileNameTo}) async {
+Future<({String? path, String? name, String? error})> selectFileAsync({
+  required List<String>? allowedExtensions,
+  required String? changeFileNameTo,
+  int? maxSize,
+}) async {
   final file = await FilePicker.platform.pickFiles(
     type: allowedExtensions == null ? FileType.any : FileType.custom,
     allowedExtensions: allowedExtensions,
@@ -22,8 +25,11 @@ Future<({String? path, String? name, String? error})> selectFileAsync(
   if (changeFileNameTo != null) {
     final fileName = file?.files.first.name;
     final fileExtension = fileName?.split('.').last;
+    final dir = xpath.dirname(file!.files.first.path!);
+
     final newFileName = '$changeFileNameTo.$fileExtension';
-    return (path: file?.files.first.path, name: newFileName, error: null);
+    final newFile = XFile(xpath.join(dir, newFileName));
+    return (path: newFile.path, name: newFile.name, error: null);
   }
   return (
     path: file?.files.first.path,