+++ lib/src/functions/select_image_camera.dart
@@ -1,13 +1,20 @@
 import 'dart:io';
 
 import 'package:image_picker/image_picker.dart';
+import 'package:path/path.dart' as xpath;
 
 /// select image from camera
 /// [maxSize] is in MB
-Future<({String? path, String? name, String? error})> selectImageCameraAsync(
-    {int? maxSize, String? changeFileNameTo}) async {
+Future<({String? path, String? name, String? error})> selectImageCameraAsync({
+  required String? changeFileNameTo,
+  required int? imageQuality,
+  int? maxSize,
+}) async {
   final picker = ImagePicker();
-  final image = await picker.pickImage(source: ImageSource.camera);
+  var image = await picker.pickImage(
+    source: ImageSource.camera,
+    imageQuality: imageQuality,
+  );
   if (image == null) return (path: null, name: null, error: null);
   if (maxSize != null) {
     final imageBytes = await image.readAsBytes();
@@ -20,7 +27,11 @@ Future<({String? path, String? name, String? error})> selectImageCameraAsync(
   if (changeFileNameTo != null) {
     final fileName = image.name;
     final fileExtension = fileName.split('.').last;
+    final dir = xpath.dirname(image.path);
+
     final newFileName = '$changeFileNameTo.$fileExtension';
+    image = XFile(xpath.join(dir, newFileName));
+
     return (path: image.path, name: newFileName, error: null);
   }
 
@@ -29,11 +40,15 @@ Future<({String? path, String? name, String? error})> selectImageCameraAsync(
 
 /// select image from gallery
 /// [maxSize] is in MB
-Future<({String? path, String? name, String? error})> selectImageGalleryAsync(
-    {int? maxSize, String? changeFileNameTo}) async {
+Future<({String? path, String? name, String? error})> selectImageGalleryAsync({
+  required String? changeFileNameTo,
+  required int? imageQuality,
+  int? maxSize,
+}) async {
   final picker = ImagePicker();
-  final image = await picker.pickImage(
+  var image = await picker.pickImage(
     source: ImageSource.gallery,
+    imageQuality: imageQuality,
   );
   if (image == null) return (path: null, name: null, error: null);
 
@@ -48,6 +63,10 @@ Future<({String? path, String? name, String? error})> selectImageGalleryAsync(
     final fileName = image.name;
     final fileExtension = fileName.split('.').last;
     final newFileName = '$changeFileNameTo.$fileExtension';
+    final dir = xpath.dirname(image.path);
+
+    image = XFile(xpath.join(dir, newFileName));
+
     return (path: image.path, name: newFileName, error: null);
   }