+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.0
+
+* Stable version.
+* solve file name not change and can't display it
+* add image quality for image picker support
+  
 ## 0.0.2
 
 * init package.